product

+++ LARAVEL/database/migrations/2024_06_30_132637_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();  // Tạo cột id với auto_increment
             $table->string('name', 50);
-            $table->float('price');
+            $table->unsignedBigInteger('price');
             $table->string('image', 255)->nullable();
             $table->text('description')->nullable();
             $table->string('status')->nullable();
